# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for the Gym Pulse workout tracking app.",
        "title": "Gym Pulse API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/exercises/history": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns, for each comma-separated exercise id, the completed sets from the most recent day that exercise was performed — powers \"last time you did X\" hints.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "logs"
                ],
                "summary": "Most recent completed sets per exercise",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Comma-separated exercise UUIDs",
                        "name": "ids",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.ExerciseHistory"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/exercises/records": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns each comma-separated exercise's heaviest weight and best estimated 1RM (Epley), from completed weighted sets.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "logs"
                ],
                "summary": "All-time records per exercise",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Comma-separated exercise UUIDs",
                        "name": "ids",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.ExerciseRecord"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/logs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all day logs in the week containing the given Monday date.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "logs"
                ],
                "summary": "List day logs for a week",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Week start date (Monday) in YYYY-MM-DD format",
                        "name": "week",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.DayLogSummary"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Logs a workout session for the given date. Only one log per date is allowed.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "logs"
                ],
                "summary": "Create a day log",
                "parameters": [
                    {
                        "description": "Day log payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateDayLogRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.DayLog"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "A log already exists for this date",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/logs/{date}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the full day log for a specific date including template and overrides.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "logs"
                ],
                "summary": "Get a day log",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Date in YYYY-MM-DD format",
                        "name": "date",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.DayLog"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Partially updates a day log. Omitted detail is preserved; explicit arrays replace or clear it. Replacing the workout clears omitted exercise detail.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "logs"
                ],
                "summary": "Update a day log",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Date in YYYY-MM-DD format",
                        "name": "date",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateDayLogRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.DayLog"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Permanently deletes the day log for the given date.",
                "tags": [
                    "logs"
                ],
                "summary": "Delete a day log",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Date in YYYY-MM-DD format",
                        "name": "date",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.CreateDayLogRequest": {
            "type": "object",
            "required": [
                "date",
                "subtype_id",
                "type_id"
            ],
            "properties": {
                "date": {
                    "type": "string"
                },
                "overrides": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.CreateOverrideRequest"
                    }
                },
                "session_notes": {
                    "type": "string"
                },
                "set_logs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.CreateSetLogRequest"
                    }
                },
                "subtype_id": {
                    "type": "string"
                },
                "template_id": {
                    "type": "string"
                },
                "type_id": {
                    "type": "string"
                }
            }
        },
...
        "model.CreateOverrideRequest": {
            "type": "object",
            "required": [
                "exercise_id"
            ],
            "properties": {
                "actual_reps": {
                    "type": "integer"
                },
                "actual_sets": {
                    "type": "integer"
                },
                "actual_weight": {
                    "type": "number"
                },
                "exercise_id": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "skipped": {
                    "type": "boolean"
                }
            }
        },
...
        "model.CreateSetLogRequest": {
            "type": "object",
            "required": [
                "exercise_id",
                "set_index"
            ],
            "properties": {
                "actual_reps": {
                    "type": "integer"
                },
                "actual_weight": {
                    "type": "number"
                },
                "completed": {
                    "type": "boolean"
                },
                "duration_seconds": {
                    "type": "integer"
                },
                "exercise_id": {
                    "type": "string"
                },
                "set_index": {
                    "type": "integer",
                    "minimum": 1
                },
                "target_reps": {
                    "type": "integer"
                },
                "target_weight": {
                    "type": "number"
                }
            }
        },
...
        "model.DayLog": {
            "type": "object",
            "required": [
                "subtype_id",
                "type_id"
            ],
            "properties": {
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "logged_at": {
                    "type": "string"
                },
                "overrides": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ExerciseOverride"
                    }
                },
                "session_notes": {
                    "type": "string"
                },
                "set_logs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SetLog"
                    }
                },
                "subtype_id": {
                    "type": "string"
                },
                "template": {
                    "$ref": "#/definitions/model.WorkoutTemplate"
                },
                "template_id": {
                    "type": "string"
                },
                "template_name": {
                    "type": "string"
                },
                "type_id": {
                    "type": "string"
                }
            }
        },
        "model.DayLogSummary": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "logged_at": {
                    "type": "string"
                },
                "session_notes": {
                    "type": "string"
                },
                "subtype_id": {
                    "type": "string"
                },
                "template_id": {
                    "type": "string"
                },
                "template_name": {
                    "type": "string"
                },
                "type_id": {
                    "type": "string"
                }
            }
        },
...
        "model.Exercise": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "catalog_id": {
                    "type": "string"
                },
                "duration_minutes": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "intensity": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 1
                },
                "notes": {
                    "type": "string"
                },
                "reps": {
                    "type": "integer"
                },
                "rest_seconds": {
                    "type": "integer"
                },
                "sets": {
                    "type": "integer"
                },
                "sort_order": {
                    "type": "integer"
                },
                "weight": {
                    "type": "number"
                }
            }
        },
        "model.ExerciseHistory": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "exercise_id": {
                    "type": "string"
                },
                "sets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SetLog"
                    }
                }
            }
        },
        "model.ExerciseOverride": {
            "type": "object",
            "required": [
                "exercise_id"
            ],
            "properties": {
                "actual_reps": {
                    "type": "integer"
                },
                "actual_sets": {
                    "type": "integer"
                },
                "actual_weight": {
                    "type": "number"
                },
                "exercise_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "skipped": {
                    "type": "boolean"
                }
            }
        },
        "model.ExerciseRecord": {
            "type": "object",
            "properties": {
                "best_e1rm": {
                    "type": "number"
                },
                "e1rm_date": {
                    "type": "string"
                },
                "e1rm_reps": {
                    "type": "integer"
                },
                "e1rm_weight": {
                    "type": "number"
                },
                "exercise_id": {
                    "type": "string"
                },
                "max_weight": {
                    "type": "number"
                },
                "max_weight_date": {
                    "type": "string"
                },
                "max_weight_reps": {
                    "type": "integer"
                }
            }
        },
...
        "model.SetLog": {
            "type": "object",
            "required": [
                "exercise_id",
                "set_index"
            ],
            "properties": {
                "actual_reps": {
                    "type": "integer"
                },
                "actual_weight": {
                    "type": "number"
                },
                "completed": {
                    "type": "boolean"
                },
                "duration_seconds": {
                    "type": "integer"
                },
                "exercise_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "set_index": {
                    "type": "integer",
                    "minimum": 1
                },
                "target_reps": {
                    "type": "integer"
                },
                "target_weight": {
                    "type": "number"
                }
            }
        },
...
        "model.UpdateDayLogRequest": {
            "type": "object",
            "properties": {
                "overrides": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.CreateOverrideRequest"
                    }
                },
                "session_notes": {
                    "type": "string"
                },
                "set_logs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.CreateSetLogRequest"
                    }
                },
                "subtype_id": {
                    "type": "string"
                },
                "template_id": {
                    "type": "string"
                },
                "type_id": {
                    "type": "string"
                }
            }
        },
...
        "model.WorkoutTemplate": {
            "type": "object",
            "required": [
                "name",
                "subtype_id",
                "type_id"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "exercises": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Exercise"
                    }
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                },
                "subtype_id": {
                    "type": "string"
                },
                "type_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Bearer token — prefix value with \"Bearer \"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}